don't try to include files that don't exist
fixes CNVS-24662

This should be an edge case, but we should not try to add files that do
not exist to either the ePub book or the unsupported files directory.

test plan:
- Have a course with assignments (or another canvas object) that
  reference files that do not exist as part of the course.
- Export the course as an ePub.
- Observe that the ePub export completes.
- Observe that the messaging in the ePub is similar to when the file is
  not supported.

Reviewed-on: https://gerrit.instructure.com/66233
Tested-by: Jenkins
QA-Review: Michael Hargiss <mhargiss@instructure.com>
Product-Review: John Corrigan <jcorrigan@instructure.com>

+++ lib/cc/exporter/epub/files_directory.rb
@@ -7,6 +7,7 @@ module CC::Exporter::Epub
 
     def add_files
       files.each do |file_data|
+        next unless file_data[:exists]
         File.open(file_data[:path_to_file]) do |file|
           zip_file.add(file_data[:local_path], file)
         end
